Refactoring: HGNs msmt-module layout
- the "constructor" functionality is now moved from a centralized
place (measurement-plane.go) to each measurement module
