+++ measurement-plane/udp-throughput/client.go
@@ -1,0 +1,1 @@
+package udpThroughput
